restore state of deploy staging branch after build
Run git restore on deployment staging after a build, once the build files have been copied out.
This eliminates potential errors from intermediary files generated as part of the build process that are normally not gitignored

In this case, that's everything in docs/api
All of these files are generated by the api docs plugin, but they are not "true" build files, so they (arguably) count as content.
I may consider adding these files to gitignore as well, since they shouldn't be edited directly. We'll see.